Fix issue with redis called on client
